[VirtualInput] Move interpolation code into it's own class
