+++ src/main/java/com/minecrafttas/mctcommon/events/EventListenerRegistry.java
@@ -252,7 +252,9 @@ public class EventListenerRegistry {
 					toThrow = null; // Reset toThrow as the correct method was found
 					method.setAccessible(true);
 					try {
-						returnValue = method.invoke(eventListener, eventParams); // Call the method
+						Object newReturnValue = method.invoke(eventListener, eventParams); // Call the method
+						if (newReturnValue != null)
+							returnValue = newReturnValue;
 					} catch (IllegalAccessException | InvocationTargetException e) {
 						throw new EventException(eventClass, e);
 					} catch (IllegalArgumentException e) {
